whitelisting for cookie and noncookie auth

+++ src/main/scala/org/broadinstitute/dsde/workbench/leonardo/api/LeoRoutes.scala
@@ -10,18 +10,25 @@ import akka.http.scaladsl.server.directives.{DebuggingDirectives, LogEntry, Logg
 import akka.http.scaladsl.server._
 import akka.stream.Materializer
 import akka.stream.scaladsl._
+import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.LazyLogging
+import net.ceedubs.ficus.Ficus._
 import org.broadinstitute.dsde.workbench.leonardo.config.SwaggerConfig
 import org.broadinstitute.dsde.workbench.leonardo.errorReportSource
 import org.broadinstitute.dsde.workbench.leonardo.model.{ClusterName, ClusterRequest, GoogleProject, LeoException}
 import org.broadinstitute.dsde.workbench.leonardo.model.LeonardoJsonSupport._
 import org.broadinstitute.dsde.workbench.leonardo.service.{LeonardoService, ProxyService}
 import org.broadinstitute.dsde.workbench.model.ErrorReportJsonSupport._
-import org.broadinstitute.dsde.workbench.model.{ErrorReport, WorkbenchExceptionWithErrorReport}
+import org.broadinstitute.dsde.workbench.model.{ErrorReport, WorkbenchEmail, WorkbenchExceptionWithErrorReport}
 
 import scala.concurrent.{ExecutionContext, Future}
 
+case class AuthorizationError(email: String) extends LeoException(s"'$email' is unauthorized", StatusCodes.Unauthorized)
+
 abstract class LeoRoutes(val leonardoService: LeonardoService, val proxyService: ProxyService, val swaggerConfig: SwaggerConfig)(implicit val system: ActorSystem, val materializer: Materializer, val executionContext: ExecutionContext) extends LazyLogging with ProxyRoutes with SwaggerRoutes with UserInfoDirectives {
+
+  private val config = ConfigFactory.parseResources("leonardo.conf").withFallback(ConfigFactory.load())
+  private val whitelist = config.as[(List[String])]("whitelist")
 
   def unauthedRoutes: Route =
     path("ping") {
@@ -35,44 +42,45 @@ abstract class LeoRoutes(val leonardoService: LeonardoService, val proxyService:
     }
 
   def leoRoutes: Route =
-    path("cluster" / Segment / Segment) { (googleProject, clusterName) =>
-      requireUserInfo { userInfo =>
-        put {
-          entity(as[ClusterRequest]) { cluster =>
-            complete {
-              leonardoService.createCluster(GoogleProject(googleProject), ClusterName(clusterName), cluster).map { cluster =>
-                StatusCodes.OK -> cluster
+    //getUserEmail() { (userEmail, whitelisted) =>
+    requireUserInfo { userInfo =>
+      if (whitelist.contains(userInfo.userEmail.value)) {
+        path("cluster" / Segment / Segment) { (googleProject, clusterName) =>
+          put {
+            entity(as[ClusterRequest]) { cluster =>
+              complete {
+                leonardoService.createCluster(GoogleProject(googleProject), ClusterName(clusterName), cluster).map { cluster =>
+                  StatusCodes.OK -> cluster
+                }
+              }
+            }
+          } ~
+            get {
+              complete {
+                leonardoService.getActiveClusterDetails(GoogleProject(googleProject), ClusterName(clusterName)).map { clusterDetails =>
+                  StatusCodes.OK -> clusterDetails
+                }
+              }
+            } ~
+            delete {
+              complete {
+                leonardoService.deleteCluster(GoogleProject(googleProject), ClusterName(clusterName)).map { _ =>
+                  StatusCodes.Accepted
+                }
+              }
+            }
+        } ~
+          path("clusters") {
+            parameterMap { params =>
+              complete {
+                leonardoService.listClusters(params).map { clusters =>
+                  StatusCodes.OK -> clusters
+                }
               }
             }
           }
-        } ~
-        get {
-          complete {
-            leonardoService.getActiveClusterDetails(GoogleProject(googleProject), ClusterName(clusterName)).map { clusterDetails =>
-              StatusCodes.OK -> clusterDetails
-            }
-          }
-        } ~
-        delete {
-          complete {
-            leonardoService.deleteCluster(GoogleProject(googleProject), ClusterName(clusterName)).map { _ =>
-              StatusCodes.Accepted
-            }
-          }
-        }
-      }
-  } ~
-  path("clusters") {
-    requireUserInfo { userInfo =>
-      parameterMap { params =>
-        complete {
-          leonardoService.listClusters(params).map { clusters =>
-            StatusCodes.OK -> clusters
-          }
-        }
-      }
+      } else throw AuthorizationError(userInfo.userEmail.value)
     }
-  }
 
   def route: Route = (logRequestResult & handleExceptions(myExceptionHandler) & handleRejections(rejectionHandler)) {
     swaggerRoutes ~ unauthedRoutes ~ proxyRoutes ~
